This patch is as a result of feedback of an audit of opencryptoki, performed in order to include opencryptoki into Ubuntu main, as a build dependency for ecryptfs-utils.  * https://bugs.launchpad.net/bugs/247593
In this patch, I attempted to find all hard coded values of 2048 that
pertained to paths.  I created #define'd a MAX_PATH value to 4096 and
changed all references I could find to that.

Note that this is entirely untested, though assuming it compiles, it
should be a rather simple (though comprehensive) set of changes.

Dustin Kirkland <kirkland@canonical.com>