Process pkcsslotd.in at build time instead of configure time
This patch moves the processing of misc/pkcsslotd.in to use
a sed script at build time instead of relying on configure
to process it at configure time. This allows us to be more
flexible when building to non-standard directories or other
odd scenarios. Also fixed to use standard directory variables.

Signed-off-by: Klaus Heinrich Kiwi <klausk@linux.vnet.ibm.com>