Process pkcs_slot.in at build time instead of configure time
This patch moves the processing of usr/sbin/pkcs_slot/pkcs_slot.in
to use a sed script at build time instead of relying on configure
to process it at configure time.

Also use the right autoconf macros to try to identify where
(and what) utilities to use.

In addition, try to use standard directory variables.

Signed-off-by: Klaus Heinrich Kiwi <klausk@linux.vnet.ibm.com>