AEP token: use compiler flags (C macro) to resolve directories
This patch moves tok_struct.h.in to tok_stuct.h in the aep token
directory as we won't be relying on configure to process this
file anymore.

Instead, use CONFIG_PATH macro that should be defined by at
build time by the compiler (and underlying build system).

Use #warnings and default values in case this is not explicitly
defined.

Signed-off-by: Klaus Heinrich Kiwi <klausk@linux.vnet.ibm.com>