don't set TERM signal explicitely when calling killproc
The TERM signal is sent by default when the killproc function is called.
Also the pid file created by the pkcsslotd executable is removed by the
initscript (resp. by the killproc function) only when no signal is set.
Other solution how to remove the pid file is to remove it directly by the
executable.

Signed-off-by: Dan Horák <dan@danny.cz>