soft token mechanism list issues
It looks like the soft token is reporting that it does SHA256 in its
mechanism list, although its not implemented in the code. A
patch to fix that follows.

Also I noticed that all the soft token mechanisms report that
they're implemented in hardware by returning the CKF_HW
flag in their mechanism info flags.  I can also submit a
patch to correct that as well...

Thanks,
Kent

Signed-off-by: Kent Yoder <key@linux.vnet.ibm.com>