automate testing
This patch enables the make installcheck target to do an automated test run against opencryptoki.  In order to make installcheck you must be root, and make will fail if the token is already initialized.  To allow a test run to succeed for a $TOKEN, you need to first run rm -rf @localstatedir@/opencryptoki/$TOKEN/*.  Requiring this ensures that a user cannot blow away an in-use token accidentally.  make installcheck will kill a running pkcsslotd.

This patch modifies
- testcases/Makefile.am
o add the installcheck target
o pre-process init_token.sh.in
- testcases/init_token.sh -> testcases/init_token.sh.in
- testcases/init_token.sh.in
o have the build fill-in the full path to pkcsconf
o return an error code to the command line if pkcsconf fails
-  testcases/ock_tests.sh.in:
o put a list of tests to run at the top
o process spaces in the output of $PKCONF parsing
o call init_token.sh for each slot we test and fail if it fails

Thanks,
Kent

Signed-off-by: Kent Yoder <key@linux.vnet.ibm.com>