add digest testing script for oc-digest
This patch adds testcases/oc-digest/digest_test.sh, which gives a flexible way to test new digests against opencryptoki.  It also modifies testcases/oc-digest.c to return a non-zero return code to the command line when it fails.  This enables the script to detect when opencryptoki doesn't support a specific digest on the slot being tested.

The script:
1. Checks that the algorithm is supported in the slot being tested
2. Generates (from /dev/urandom) a set of files to test based on the block size of the supported algorithm
3. Hashes those files using oc-digest and validates the output
4. Deletes the test files

This initial version of the script supports testing md5, sha1, sha256, sha384 and sha512.

Thanks,
Kent

Signed-off-by: Kent Yoder <key@linux.vnet.ibm.com>