Unextractable keys should not be revealed off the token
As reported by PKCS#11 standard on the "Security considerations" section,
unextractbale keys cannot be revealed off the token neither in plaintext
nor in encrypted form.

The attached patch prevents the opencryptoki's software token to reveal
the value of a key in plaintext when its CKA_EXTRACTABLE attribute is
set to FALSE (even if the key is not a CKA_SENSITIVE one).

Patch by bugant <bugant@gmail.com>