Added token_keysize to the token_specific_struct.
When token_keysize is set, that indicates we are using a secure key
token and the keysize for the secure key.
Otherwise, it is set to 0, indicating this is not a secure key token and
thus no specific keysize required.
Thus token_keysize can also be used to indicate a secure key token.

This was needed so that symmetric key gens could know when to save the
token generated key to a CKA_VALUE or a CKA_IBM_OPAQUE.
Fixed the symmetric key gens (aes, des, des3).
Also, this fixes master key handling which was broken in cca token because
we were building CKA_IBM_OPAQUE in token and passing back zeroes for the
generated key.

Signed-off-by: Joy Latten <jmlatten@linux.vnet.ibm.com>
Reviewed-by: Debora Babb <dvelarde@us.ibm.com>
Reviewed-by: Marcelo Cerri <mhcerri@linux.vnet.ibm.com>