test: fix key wrapping test case
Wrapped key length should only be provided as an attribute when using
the CKM_RSA_X_509 mechanism. CKM_RSA_PKCS holds the key length in the
wrapped data and providing it as an attribute causes ICSF to return an
error.

Also, this commit fixes how information about wrong key length values
are shown.

Signed-off-by: Marcelo Cerri <mhcerri@linux.vnet.ibm.com>
Reviewed-by: Joy Latten