Patch[4/4]: Part 2 of slot-info via socket.
This patch enables all the non-changing slot information to be
transmitted over  the sockets, rather via shared memory.  By doing this,
the members of the pkcs11 group can no longer override the locations of
the backend library .so files, and so can no longer execute arbitary
code with the preivileges of other pkcs11 group members.

Signed-off-by: Ross McIlroy <rmcilroy@google.com>