SOFT: Fix soft token implementation of digest functions
The way that the SOFT token was implemented to deal with digest
functions, mostly SHA, was to call OpenSSL to deal with it, but OpenSSL
would try to rely on the implementation of the engine that is loaded. So
whenever we have the ibmpkcs11 engine loaded, OpenSSL would call the
ibmpkcs11, which would again call the SOFT token implementation, which
would again call OpenSSL, leading to a infinite loop. So instead of
calling OpenSSL as it was done, we call OpenSSL digest primitive functions
directly.

Signed-off-by: Eduardo Barretto <ebarretto@linux.vnet.ibm.com>