Remove outdated documents
OpenCryptoki-HOWTO.pdf is from 2005, really outdated, and we shouldn't have a
pdf in the package source. Another file will be written to take its place.
README-IBM_CCA_users was needed when pkcsconf didn't support 64 bits, so there
was a pkcsconf64 tool to help on that. But now this pkcsconf64 doesn't exist
anymore and pkcsconf works for 32 and 64bits.

Signed-off-by: Eduardo Barretto <ebarretto@linux.vnet.ibm.com>