SOFT: Fix for the case where an engine is loaded on Openssl
If OpenSSL is loading an engine, we might end up having an infinite loop, since
the SOFT token doesn't implement RSA encryption/decryption and calls OpenSSL to
solve it. So to fix it, we check if an engine is loaded and, if yes, we set RSA
methods to the default RSA methods.

Signed-off-by: Eduardo Barretto <ebarretto@linux.vnet.ibm.com>