EP11: Make CKA_SENSITIVE default to TRUE for secret keys
When option FORCE_SENSITIVE is specified in the ep11 config
file, then for secret keys (CKA_CLASS = CKO_SECRET_KEY) the
attribute CKA_SENSITIVE is set to CK_TRUE if it is not explicitly
specified in the template.

This applies to C_GenerateKey, C_GenerateKeyPair, C_UnwrapKey,
and C_DeriveKey.

Signed-off-by: Ingo Franzki <ifranzki@linux.vnet.ibm.com>