EP11: Add functions to query control points of EP11 cards
Add a set of functions to query the control points of EP11
cards. If the ep11 config file uses APQN_ANY, then sysfs is
used to determine all available EP11 cards.
If an APQN_WHIELIST is used, then only those cards in the
white list are used.

When the cards have different control points settings, a
warning message is issued to syslog, and the minimum
control point setting of all cards is returned.

Signed-off-by: Ingo Franzki <ifranzki@linux.vnet.ibm.com>