EP11: Add function to parse CP-filter config file
The list of mechanisms returned by C_GetMechanismList is
filtered using the control point settings of the used crypto
adapters.The EP11 CP-filter config file is used to
associate certain control points with mechanisms that are
dependent on these control points. The default CP-filter
config file is 'ep11cpfilter.conf' located in the same
directory as this EP11 token configuration file.
You can optionally specify the name and/or location of the
CP-filter file in the EP 11 Token Config file with keyword
CPFILTER:
      CPFILTER /etc/opencryptoki/ep11cpfilter.conf

The syntax of the CP-Filter file is as follows:
      cp: mech1, mech2, ...

Signed-off-by: Ingo Franzki <ifranzki@linux.vnet.ibm.com>