EP11: pkcsep11_*: Prompt for pins instead of using environment variables
For security reasons it is not a good idea to use environment variables
to pass the PKCS#11 pins to the tools. This patch changes the two EP11
related tools pkcsep11_migrate and pkcsep11_session to prompt for the
pins instead of using environment variables.

All other tools provided as part of the Opencryptoki package already
prompt for pins, so this patch makes the EP11 related tools to behave
the same way as the other tools.

Signed-off-by: Ingo Franzki <ifranzki@linux.vnet.ibm.com>