Fix testcase after max number of slots managed was increased
The multi instance token support patch series also increased the
maximum number of tokens managed (NUMBER_SLOTS_MANAGED in slotmgr.h)
from 6 to 1024. The testcase gen_purpose passes a slot-id that
is supposed to be out of range to functions C_GetSlotInfo, C_GetTokenInfo,
C_GetMechanismList, and C_InitToken and expects CKR_SLOT_ID_INVALID
to be returned. With the increased number of managed slots, the
slot-id used by the testcase (i.e,. 999) is no longer out of range.

This commit changes the testcase to use slot-id 9999, so that it is
out of range again.

Signed-off-by: Ingo Franzki <ifranzki@linux.vnet.ibm.com>