Update token structs for ECC support in icatoken.
The ecdh_derive function is new in all tokens, so we must
add a function entry for it in all token specific structs
and initialize it with NULL. Also, a function prototype
must be added to the common token structs. ECDH mechanism
related structs and defines are added to pkcs11types.h.

Signed-off-by: Joerg Schmidbauer <jschmidb@de.ibm.com>