Fix copy object when template is null
When some application call C_CopyObject and passes a null template and a
count of 0 then we should be able to continue and execute the copy, but
this was failing because of some wrong checks.
As a consequence I've also added a testcase to copyobjects to try to
copy an object when we pass a null template and check if the new object
matches the old object.

Signed-off-by: Eduardo Barretto <ebarretto@linux.vnet.ibm.com>