ICA: Fix token_specific_rsa_x509_verify() function
A break is missing in a loop to find the beginning of the
output data. This causes C_Verify to fail with
CKR_SIGNATURE_INVALID when mechanism CKM_RSA_X_509 is used.

This function was introduced in 2012 with commit
e745e536f70a534dd343799967e21bc5d13d9e2d and could
never have worked since then.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>