Fix a bug in sha512_hmac_sign()
A wrong size during manual HMAC calculation causes that the
returned signature is wrong. This happens for the ICA token,
since it does not provide a token specific HMAC_Sign function.

Also remove an unnecessary second check for the token specific
HMAC-Sign function. This check is already done a few lines above and
is not needed again. If we reach there, there is no token specific
HMAC-Sign function.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>