EP11: Performance enhancement for mechanism validation
Every SC_xxx function that gets passed a mechanism calls
function valid_mech() to validate if the mechanism is supported.
This currently causes a call to the EP11 library (m_GetMechanismInfo)
and thus a call to the EP11 crypto adapter (which is quite time
consuming).

This patch removes the call to m_GetMechanismInfo() in valid_mech()
and just checks the mechanism against the banned mechanism list, as
well as against the control point filter. If an SC_xxx function is
called with a mechanism that is not supported by the EP11 library and/or
the EP11 crypto adapter, the appropriate EP11 library call will return
CKR_MECHANISM_INVALID anyway.

Function SC_GetMechanismInfo will still call m_GetMechanismInfo()
but will also check for banned or filtered out mechanisms.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>