EP11: Use m_Encrypt/DecryptSingle for single part encrypt/decrypt calls
To optimize performance for single part encrypt/decrypt calls the init
call is not passed through the EP11 library as long as there is
no multi-part operation. In case of a single part operation,
the m_Encrypt/DecryptSingle function of the EP11 library is used.
In case of a multi part operation, the init call is deferred up to
the first update call.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>
Signed-off-by: Ingo Tuchscherer <ingo.tuchscherer@linux.ibm.com>