Fix bug with master key encryption with FIPS enabled libica
When running with a FIPS enabled libica, the ICA token fails to
initialize, because the 3DES key derived from the user or SO pin is
considered invalid because the first and the third part of the
3DES key is the same.

For clear key tokens, the token specific 3DES-CBC function is
used for the master key encryption. In case of the ICA token,
the ICA token specific 3DES-CBC function fails, because libica
rejects the key when compiled with FIPS support. This leads to
an error during token initialization.

Instead of using the token specific 3DES-CBC function, the code
now always falls back to the (OpenSSL) based software encryption
function, as it is also done for secure key tokens.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>