Reject C_Digestkey for secure key tokens
digest_mgr_digest_key() uses attribute CKA_VALUE as the key to be
digested and then calls digest_mgr_digest_update() to add the key
to the digest. This is wrong for secure key tokens, since secure
keys that can not be digested that way, because attribute
CKA_VALUE does not contain the key to be digested.

Return CKR_FUNKTION_NOT_SUPPORTED in case digest_mgr_digest_key()
is called for a secure key token.

This fixes issue #16:
https://github.com/opencryptoki/opencryptoki/issues/16

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>