Fix C_GetSlotList behavior
"1. If pSlotList is NULL_PTR, then all that C_GetSlotList
does is return (in *pulCount) the number of slots, without
actually  returning a list of slots. The contents of the
buffer pointed to by pulCount on entry to C_GetSlotList
has no meaning in this case, and the call returns the value
CKR_OK.
2. If pSlotList is not NULL_PTR, then *pulCount must contain
the size (in terms of CK_SLOT_ID elements) of the buffer
pointed to by pSlotList. If that buffer is large enough to
hold the list of slots, then the list is returned in it, and
CKR_OK is returned. If not, then the call to C_GetSlotList
returns the value CKR_BUFFER_TOO_SMALL. In either case, the
value *pulCount is set to hold the number of slots."
[pkcs11]

Signed-off-by: Patrick Steuer <patrick.steuer@de.ibm.com>