Add is_secure_key_token() function
For better code readability, change constructs like
    if (token_specific.token_keysize)
to use function is_secure_key_token(). This is no
change in the logic, but makes the code easier to
understand.

Secure key tokens specify a token key size of > 0, so
this can be used as an indication for a secure key
token.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>