CCA, EP11, ICA: debug instead of error trace in case of unsupported mechanism
On invocation of token_specific_get_mechanism_info() the CCA, EP11 and ICA
token trace an ERROR line at unsupported mechanism. This function is called
to check, if a mechanism is valid or not and thus should not emit an error
trace if just the mechanism is unknown to the token. So reworked the trace
statement to emit a DEBUG trace line instead of ERROR.

Signed-off-by: Harald Freudenberger <freude@linux.ibm.com>