CCA Token: fix misleading trace statement
The ccatok_hmac function traced a line
   ccatok_hmac: key length is %lu
but this is the byte length of the cca secure key
token NOT the size of the hmac key. This is confusing
and does not carry any useful information.

Reworked this statement to show the hmac key size
in bytes instead which has some use for debugging.

Signed-off-by: Harald Freudenberger <freude@linux.ibm.com>