Revert: Compare check for token min and max keysize correction
The compare check for token min and max keysize as changed by
commit 4cdd5038e732ab895fae64a32face784fee000f8 is not correct.

Key size units as reported by C_GetMechanismInfo are in bits
or in bytes dependent on the mechanism. Leave it up to the
caller of check_supp_keysize() to use the correct unit for the
mechanism specified. That way function check_supp_keysize() is
mechanism independent.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>