EP11: add support for querying the EP11 lib and firmware version
Sometimes the EP11 token must adjust the behavior of the EP11
host library when its behavior is not exactly as defined by the
PKCS#11 standard. This adjustments are mostly EP11 host library
or even firmware version dependent. Also some new features
are only available/working with a certain host library or
firmware version.

This patch adds code to query the versions of the host library,
the firmware API level and the firmware version, so that this
version information can be used later on.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>