pkcsslotd: adapt pid file creation on demonize for systemd
The current implementation creates the pid file of the
pkcsslotd demon as one of the last steps of the slot
demons startup. It's done after the parent has spawned
away the child which will become the demon when this child
is nearly done with initialization. At this time the
parent has already terminated.

However, systemd expects to find the pid file when the
parent terminates. So here is now a patch which moves
the creation of the pid file into the parent code. Now
the pid file is created after the child has been forked
in the parent process short before it calls exit().

Signed-off-by: Harald Freudenberger <freude@linux.ibm.com>