EP11: Add helper function to check for required firmware version
To allow easy checking if the configured APQNs support a certain
minimum firmware version, the newly added helper function can be
used.

To check the versions for multiple card types (CEX4P, CEX5P, ...)
call this function for each card type with the minimum versions for
that card type. Only when this function returns TRUE for all card
types, then it can be assumed that all configured APQNs support the
required minimum level.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>