Fix race condition with save_token_data/load_token_data()
The lock is released before the file is closed when saving the
token data. This leaves a very small window where the file is not
closed, but other processes may read the file. In this case the
file may appear with a size of 0 bytes, thus loading the token
data fails.

Fixes: https://github.com/opencryptoki/opencryptoki/issues/218

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>