pkcsep11_session: Allow already closed sessions
When multiple VHSM-Session objects are left over to be
logged out by the pkcsep11_session tool, it may happen
that all or some of them contain the exact same session
pin blob. Thus when logging out one of them, all others
will receive CKR_SESSION_CLOSED when trying to log them
out. This is not an error, since the session is already
logged out in that case.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>