EP11: Fix target_list passing for EP11-session logon/logoff
The target list is passed incorrectly to the handle_all_ep11_cards()
function. Dependent on memory content, this can cause it to
fail processing all EP11 cards. This will most likely cause a
CKR_DEVICE_ERROR to be returned by C_Login when the STRICT_SESSION
and/or VHSM_MODE is enabled in the ep11tok.conf config file.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>