EP11: Support key usage restrictions for public keys
EP11 does not allow to restrict a public RSA, EC, or DSA keys using
CKA_VERIFY=FALSE or CKA_ENCRYPT=FALSE, because it can not be technically
enforced since the public key material is available in clear
(refer to chapter 2.2.6 of the EP11 structure document).
Thus, generating a key pair using C_GenerateKeyPair fails with
CKR_TEMPLATE_INCONSISTENT when the public key template contains
CKA_VERIFY=FALSE or CKA_ENCRYPT=FALSE.

With this change, the EP11 token allows to restrict public keys
with CKA_VERIFY=FALSE or CKA_ENCRYPT=FALSE. When specified, these
attributes are overridden with TRUE when passing the key generate
request to the EP11 library, but the original attribute values are
stored in the object. The key usage restriction in regards of SIGN
and ENCRYPT are enforced by the EP11 token, before the sign or
encrypt request is passed to the EP11 library.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>