Fix coverty scan findings for stat/open
A stat() before open() is not required and can cause timing
ralted errors. The code can directly react on the result of
open(). Flag O_CREAT will create the file only if it is not
existent, so we don't need to know upfront if the file exists
already or not.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>