EP11: Load versioned EP11 host library shared object
The EP11 token currently loads the un-versioned EP11 host library
shared object (i.e. libep11.so). This may cause problems in the
future, when newer versions of the EP11 host library no longer
supply the un-versioned shared object (or an un-versioned symlink
to the shared object).

Change the EP11 host library loading to first try to load a version 2
shared object (libep11.so.2). If this is not available, then try to
load a version 1 shared object (libep11.so.1), and if this is also
not available, try to load the un-versioned shared object (libep11.so).

That way the EP11 token always loads the newest available EP11 host
library version (that openCryptoki supports). At the same time we
stay backward compatible and can still work with older host library
versions. For special cases, environment variable OCK_EP11_LIBRARY
can be set to specify the name of the shared object to load.

Note: The EP11 token queries the host library version. Based on
the version some features may not be available, or may behave
differently.

Fixes: https://github.com/opencryptoki/opencryptoki/issues/243

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>