Provide an external header file with the supported EC curves
To make it easier for applications to specify the DER encoded OIDs
of the desired EC curve, provide an external, but non-PKCS11-standard
header file that contains defines for the supported EC curves. These
defines can be used as array initializers. The DER encoded OIDs
are specified as attribute CKA_ECDSA_PARAMS or CKA_EC_PARAMS with
C_GenerateKeyPair to generate an EC key pair for the desired curve.

While at it, also rename some of the static variables to match the
exact EC curve names.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>