TESTCASE: Enhance CKM_ECDH1_DERIVE tests for EP11
EP11 supports CKM_ECDH1_DERIVE with or without KDFs and shared data
dependent on the firmware version. Enhance the tests so that it skips
test test when it gets error CKR_MECHANISM_PARAM_INVALID when using the
EP11 token and a KDF other than CKD_NULL and/or shared data is used.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>