EP11: Special handling for CKM_ECDH1_DERIVE with CEX7
With CEX7 and the EP11 host library version 3 the CKM_ECDH1_DERIVE
mechanism supports the CK_ECDH1_DERIVE_PARAMS as of PKCS#11 v2.11.

If the set of crypto adapters are all CEX7 cards, then the used
API version is 3 and CK_ECDH1_DERIVE_PARAMS can be passed as is.
If older cards are part of the set of cards, then API version 2
is used, and we still need to pass the public data only as
mechanism parameter.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>