COMMON: Fix btree node free race condition
The bt_node_free function frees the node by marking it
as free by setting flag BT_FLAG_FREE and adding it to
the free list. It also calls the specified callback
function with the node value.

To avoid a race condition, it must first set the flag to
mark the node deleted (i.e. free), and then call the callback
function. The callback function might free the node value.
If the node is not yet marked as deleted when the value is
being freed, then another thread might access the node and its
value, although the value has already been freed.

This problem may for example happen when one thread runs
C_FindObjectsInit, and another thread deletes an object via
C_DestroyObject at the same time.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>