Update token pin and data store encryption for soft,ica,cca and ep11
A new config option "tokversion = <major>.<minor>" is added to
opencryptoki.conf. This new optional key-value pair tells openCryptoki
which version of a token is associated with a slot. If the token
version is not specified explicitely, openCryptoki uses a default
version of 0.0.
If new versions of soft,ica,cca and ep11 token are set to 3.12
the new pin encryption and data store format will be used.
The 3.12 format is incompatible with the old 0.0 format i.e.,
you must not use this new format with existing data stores and
openCryptoki setups. It must only be used with fresh openCryptoki
configurations. Migration tools to migrate existing setups
to the new version will be provided in the future.

Signed-off-by: Patrick Steuer <patrick.steuer@de.ibm.com>