TESTCASES: Enhance RSA-OAEP testcases
EP11 supports CKM_RSA_PKCS_OAEP with non-SHA1 hashing and MFG algorithms
dependent on the firmware version. Enhance the tests so that it skips
test test when it gets error CKR_MECHANISM_PARAM_INVALID when using the
EP11 token and an RSA-OAEP mechanism parameter with non-SHA1 hashing and/or
MGF algorithm.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>