pkcsep11_migrate: Rework to replace deprecated EP11 functions
Some EP11 host library functions used by the pkcsep11_migrate tool are
now deprecated and are replaced by equivalent functions with a different
name. Change the code to use the current function names, but fallback
to the old names if running with a very old EP11 host library version
that does not yet provide the new functions.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>