TESTCASES: Add multi instance test case
Open sessions on two different tokens within one process. Ensure that
both tokens operate independently, and that sessions of objects opened
with one token can not be used with the other one.

To test multi-instance support of a token, the Opencryptoi configuration
must be set up that way that 2 tokens of the same type (e.g. 2 EP11 tokens),
exist.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>