Simplify btree function bt_node_free
There exist 2 almost equal functions named 'bt_node_free()' and
'bt_node_free_()'. The only difference is the signature of the callback
routine used. For one the callback just gets the node value, and for the
other it gets the node value and the STDLL_TokData_t of the token.

Simplify the code by removing the duplicate function and always use
the callback function with just the value.
Use the same callback function signature for bt_destroy().

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>