Fix btree race conditions
Ensure that we always obtain the value from a node while holding the
btree lock, or from within a transaction block. Otherwise the node may
be deleted concurrently after we got the node, but before we got the
value from the node. The value of a deleted node points to other nodes
in the btree's free list. Thus, we would potentially return an invalid
value pointer.

With this changes, we either return a valid value of a node, or NULL if
it has been deleted.

Note that the value might still get deleted (i.e. freed) after we got it,
if the node is deleted concurrently (dependent on what the bt_node_free()
callback function does). The btree lock can _NOT_ protect from this situation.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>