Change bt_node_free() to return the value of the deleted node
... and not the node itself. A deleted node is chained into the free list
using the value field of the node, so a returned deleted node is of no use
for the caller. From the deleted node you can't get the deleted value.

Return the value of the deleted node instead. Depending on if a delete
callback function is specified and what it does, the returned value might
have already been freed by the callback. Nevertheless, a non-NULL value
returned still indicates that the node to delete was found and has been
deleted. To use the value after the node has been deleted, don't specify
a callback routine and free the value on your own later on.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>