Make XProcLock a truly recursive lock
The XProcLock consists of 2 locks, a file lock, for protecting the
shared memory and token directory from concurrent access by processes,
and a (recursive) mutex, for protecting the shared memory and token
directory for concurrent access by other threads of the same process.

While the mutex is created as recursive mutex, the file lock is not
recursive. Add a lock counter for the file lock, to make it recursive,
as well. The counter for the file lock is protected by the recursive
mutex, so no concurrent update of the counter can happen from other
threads.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>