Correct locking of access to token objects
Ensure the token object btrees, the shared memory, and the token directory
are in sync all the time. Keep the XProcLock until all 3, the token object
btrees, the shared memory, and the token directory have been updated.

Otherwise it might happen that a token object btree already contains a token
object, but it has not been added to the shared memory yet. This may lead to
the situation where a call to object_mgr_update_from_shm() during
C_FindObjectsInit finds the object in the betree, but not in the shared memory,
and thus deletes the object from the btree.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>