TESTCASE: Add multi threaded object lock test
Use an object in a set of usage threads, and concurrently modify
that object in a set of alter threads. Ensure that the object is
never in an inconsistent state.

To test concurrent alteration of token objects, run this testcase twice
at the same time, so that the token object is also modified by other
processes, and thus is reloaded frequently. Run the tetscase once with
options -token_obj and -keep_obj to create the key object, but do not
destroy it afterwards. For any subsequent runs, use options -token_obj,
-reuse_obj and -keep_obj to operate on the already existing object.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>