ICSF: Do not close parents LDAP connection in forked client
When a process is forked while it has open sessions with the ICSF
token, then the fork initializer running in the client process should
_NOT_ close the LDAP connections associated with the sessions during
C_Finalize inside the fork initializer.

Closing the LDAP connection in the client would also close the connection
for the parent process, since client and parent share the same socket
after the fork. However, the parent still wants to continue working with
its LDAP connection after the fork.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>