EP11: Filter out unknown/unsupported mechanisms
Change the code from using a negative list (banned mechanisms) to using
a positive list (supported mechanisms) instead, to filter out unknown
or unsupported mechanisms. Now, only those mechanisms are reported that
are supported (i.e. known) by the EP11 token code, and by the configured
set of APQNs.

This has the benefit that new mechanisms added by future EP11 crypto
card enhancements are not automatically reported as supported by the
EP11 token, although it is not clear if the EP11 token code really
supports that mechanism. With the negative list, such new mechanisms
that are not on the negative list because they have been unknown so far,
would be reported as supported. With the new positive list approach,
these mechanisms remain unsupported, until they are explicitly added
to the list of supported mechanisms.

This also helps to solve problems with mixed card version environments,
where cards of different versions are in the set of configured APQNs of
the EP11 token. Here, a newer card may support a new, so far unknown mechanism,
but older cards do not support it. For mechanisms known by the EP11 token,
it is verified that the mechanism is supported by all APQNs, otherwise the
mechanism is not reported via C_GetMechanismList. For unknown mechanisms
this can not be checked. So it can happen that a mechanism is reported that is
only supported by one APQN, but not by the others. This may lead to errors,
when that mechanism is used, or even when C_GetMechanismInfo is used to obtain
information about the mechanism (as done by pkcsconf -m). Inconsistencies can
also happen when C_GetMechanismList is called several times (e.g. once to get
the number of mechanisms, and a second time to obtain the list).
With the positive list approach, unknown mechanisms will never be reported, so
these problems can not happen.

Signed-off-by: Ingo Franzki <ifranzki@linux.ibm.com>